requirements.txt

pyspark
kafka-python
pandas
matplotlib
